Added create and update for subject and teacher entities

+++ src/main/kotlin/ua/edu/ukma/dudes/scheduleMeBaby/repository/TeacherRepository.kt
@@ -3,4 +3,4 @@ package ua.edu.ukma.dudes.scheduleMeBaby.repository
 import org.springframework.data.repository.CrudRepository
 import ua.edu.ukma.dudes.scheduleMeBaby.entity.Teacher
 
-interface TeacherRepository : CrudRepository<Teacher, Int>
+interface TeacherRepository : CrudRepository<Teacher, Long>
